% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_detect.R
\name{ci_detect}
\alias{ci_detect}
\title{Curve intervention detection - statistical identification and removal of disturbances in tree ring series}
\usage{
ci_detect(
  rwl,
  detrend.method = "None",
  nyrs = NULL,
  min.win = 9,
  max.win = 30,
  var.type = "s_bi",
  thresh = 3.29,
  dist.span = 1.25,
  max.iter = 10,
  add.recent.rwi = TRUE
)
}
\arguments{
\item{rwl}{A rwl object (read in by \code{\link[dplR]{read.rwl}}). Essentially a data.frame with columns names as series IDs and years as rownames.}

\item{detrend.method}{A character vector specifying the detrending method to use. Passes to \code{\link[dplR]{detrend}}. Default is "Mean".}

\item{nyrs}{A numeric vector that determines the flexibility of the \code{"AgeDepSpline"} or the \code{"Spline"} detrending methods. Passes to \code{\link[dplR]{detrend}}. The default is 50 years or 1/3 the series length.}

\item{min.win}{The minimum disturbance length in years (i.e., a moving window) to search for. The default is 9.}

\item{max.win}{The maximum disturbance length in years (i.e., a moving window) to search for. The default is the smallest of 30 years or 1/3 of the series length.}

\item{thresh}{The disturbance detection threshold, corresponding to the number of deviations from the robust mean. The default is 3.29, following Druckenbrod et al. 2013.}

\item{dist.span}{Parameter to determine the wiggliness of the loess splines fit to disturbances periods. Higher numbers = less wiggles. Passes to \code{\link[stats]{loess}}. The default is 1.25}

\item{max.iter}{The maximum number of iterations to run the disturbance detection and removal processes. The default is 10.}
}
\value{
A 5-element list containing the "disturbance-free" series, the disturbance index,
a data.frame containing basic data on all the detected disturbances,
a record of the disturbance detection and removal iterations, and the output from the cp_detrend() process.
}
\description{
This function implements an analysis based on the curve intervention detection techniques described by Druckenbrod et al. 2013,
Rydval et al. 2016, and Rydval et al. 2018. The modendro implementation differs from the version described by these authors. See Details for more information.
The basic motivations for performing this analysis are to identify, quantify, & remove abrupt changes in radial growth. You might use
this information to help make inferences about forest stand dynamics, insect outbreaks, or isolate a stronger climate signal, depending on your dendroecological or dendroclimatological tendencies.
}
\details{
Intervention detection is a statistical time series approach to identify and remove abrupt changes in radial growth. The "curve" part
of curve intervention detection (CID) describes the type of line fitted to each period that is identified as a disturbance.
Following Rydval et al. (2016, 2018), the implementation here uses a version of the equation presented by Warren & MacWilliam (1981), with
slightly different coefficients than the Hugershoff curve used by Rydval et al. Namely, the following values are fixed: b = 1, and d = 0.
The Warren & MacWilliam curve has a 't' coefficient which we use here. These changes improve the robustness of the fits. Fitting is done
via \code{\link[stats]{nls}}. A major difference for the \code{modendro} implementation of CID is that in the cases where the \code{\link[stats]{nls}} fits fail, a \code{\link[stats]{loess}} spline is fit
instead. See \code{\link{dist_det_rem}} for more details.

\code{\link{function}}
}
\examples{
library(dplR)
data("co021")
# before
ci_detect(co021, max.iter = 5)
}
\references{
Druckenbrod, D. L., N. Pederson, J. Rentch, and E. R. Cook. 2013. A comparison of times series approaches for dendroecological reconstructions of past canopy disturbance events. \emph{Forest Ecology and Management}, \strong{302}, 23–33.

Rydval, M., D. Druckenbrod, K. J. Anchukaitis, and R. Wilson. 2016. Detection and removal of disturbance trends in tree-ring series for dendroclimatology. \emph{Canadian Journal of Forest Research}, \strong{401}, 387–401.

Rydval, M., D. L. Druckenbrod, M. Svoboda, V. Trotsiuk, P. Janda, M. Mikoláš, V. Čada, R. Bače, M. Teodosiu, and R. Wilson. 2018. Influence of sampling and disturbance history on climatic sensitivity of temperature limited conifers. \emph{The Holocene}, \strong{28}(10), 1574-1587.

Warren, W. G., and S. L. MacWilliam. 1981. Test of a new method for removing the growth trend from dendrochronological data. \emph{Tree Ring Bulletin}, \strong{41}, 55–66.
}
