% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idPRISM}
\alias{idPRISM}
\title{PRISM climate data to accompany Perkins & Swetnam whitebark pine collection
1895-1991}
\format{
\subsection{\code{idPRISM}}{

An rwl data.frame with 1164 rows and 5 columns:
\describe{
\item{Date}{Date in \%Y-\%m-\%d format}
\item{year}{Year}
\item{month}{Month}
\item{PPT.mm}{Monthly total precipitation, in mm}
\item{Tavg.C}{Monthly average daily temperature, in °C}
...
}
}
}
\source{
\url{https://www.prism.oregonstate.edu/explorer/}
}
\usage{
idPRISM
}
\description{
PRISM climate data to accompany Perkins & Swetnam whitebark pine collection
1895-1991
}
\keyword{datasets}
