% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idPRISMgroup}
\alias{idPRISMgroup}
\title{Site-specific PRISM climate data to accompany Perkins & Swetnam whitebark pine collection
1895-1991}
\format{
\subsection{\code{idPRISMgroup}}{

An rwl data.frame with 3492 rows and 6 columns:
\describe{
\item{Date}{Date in \%Y-\%m-\%d format}
\item{year}{Year}
\item{month}{Month}
\item{PPT.mm}{Monthly total precipitation, in mm}
\item{Tavg.C}{Monthly average daily temperature, in °C}
\item{site}{Whitebark pine sampling location}
...
}
}
}
\source{
\url{https://www.prism.oregonstate.edu/explorer/}
}
\usage{
idPRISMgroup
}
\description{
Site-specific PRISM climate data to accompany Perkins & Swetnam whitebark pine collection
1895-1991
}
\keyword{datasets}
