% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longer_rwl.R
\name{longer_rwl}
\alias{longer_rwl}
\title{Convert long format ring-width data to rwl-format}
\usage{
longer_rwl(df = NULL, dat.name = NULL)
}
\arguments{
\item{df}{A long-format data.frame with "year" & "series" columns and a data column, specified
by the dat.name argument}

\item{dat.name}{A character vector specifying the name of the tree ring data
(e.g., "rw", "bai.cm2", "d13C", etc.)}
}
\value{
A data.frame with series names as columns and years as rownames.
}
\description{
Simple function to convert long format tree ring data (3 columns: year, series, rw) to rwl-format
data (columns are series, rows are years). This is useful for grouping and reassubling data.
}
\examples{
library(dplR)
# Bristlecone pine tree ring collection from Campito mountain, White Mountains, CA
# Many 0 value rings in this collection
data("ca533")

ca533.long <- rwl_longer(rwl = ca533, dat.name = "rw.mm", trim = TRUE)
head(ca533.long)

# Convert it back to rwl-like format

ca533.rwl <- longer_rwl(df = ca533.long, dat.name = "rw.mm")
head(ca533.rwl)
all(ca533.rwl \%in\% ca533)

}
