% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mo024}
\alias{mo024}
\title{Post oak (QUST) tree-ring width collection from Democrat Ridge, Missouri 1992 update
D. Stahle, G. Hawks, S. Sierzchula, N. Montagu}
\format{
\subsection{\code{mo024}}{

An rwl data.frame with 373 rows and 61 columns (i.e., 61 tree-ring series spanning 373 years)
}
}
\source{
\url{https://www.ncei.noaa.gov/access/paleo-search/study/4836}
}
\usage{
mo024
}
\description{
Post oak (QUST) tree-ring width collection from Democrat Ridge, Missouri 1992 update
D. Stahle, G. Hawks, S. Sierzchula, N. Montagu
}
\keyword{datasets}
