% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_win_multi.R
\name{moving_win_multi}
\alias{moving_win_multi}
\title{Calculate multiple-length moving windows of monthly climate data}
\usage{
moving_win_multi(df, clim.var, win.lens, agg.fun = NULL)
}
\arguments{
\item{df}{a data.frame or matrix of climate data, with at least 3 columns: year, month (numeric),
and a climate variable.}

\item{clim.var}{a character vector of length = 1 that specifies the column name in \code{df} that
refers to the climate variable you wish to calculate moving windows for.}

\item{win.lens}{a numeric sequence specifying which moving windows (in months) to calculate. In
\code{\link{n_mon_corr}}, this will be 2:12.}

\item{agg.fun}{character vector of length = 1 that specifies whether to take the mean (e.g., for
temperature) or sum (e.g., for precipitation) of the moving windows. Options are
\code{c("mean","sum")}.}
}
\value{
A data.frame with the same number of rows as \code{df} - essentially it is \code{df} with the
moving windows attached with \code{\link[base]{cbind}}. For each of the moving windows, the
numeric month variable indicates the start month of the window. The length of the window is
indicated in the column name (e.g., "mean2", "mean6", "mean12", etc.).
}
\description{
This function is a workhorse function behind \code{\link{n_mon_corr}}. It very efficiently
calculates the moving window sums or means by taking lagged differences from the cumsum() of
the properly ordered climate data.
}
\examples{
# Make some fake climate data for testing and building the function
mo <- 1:12
x <- seq(-4, 4, length.out = 12)
# Give the clim.var a seasonal fluctuation, kind of like temperature outside of
# tropical latitudes
gauss.curv <- \(x) {(10/sqrt(2*pi*1.6))*exp(-((x^2)/(2*1.6^2)))}
clim.mo <- data.frame(month = mo, clim.var = gauss.curv(x))

clim.list <- vector("list", length = 12)
for (y in seq_along(clim.list)) {
  clim.y <- data.frame(year = 1951:2000, month = y)
  clim.y[,"clim.var"] <- runif(50, min = 0.1, max = 2) +
  clim.mo[clim.mo$month \%in\% y, "clim.var"]
  clim.list[[y]] <- clim.y
}

fake.clim <- do.call("rbind", clim.list)

test.output <- moving_win_multi(df = fake.clim,
clim.var = "clim.var",
win.lens = 2:12,
agg.fun = "mean")

# Make a plot of 2-, 6-, and 12-month moving window means.
library(ggplot2)

# Add a date column to make plotting easier
test.output$yr.mo <- as.Date(paste(test.output$year,
ifelse(nchar(test.output$month) == 1,
       paste0("0", test.output$month),
       test.output$month), "01",
sep = "-"), format = "\%Y-\%m-\%d")

# The continuous moving windows - note that the month on the x-axis refers to the start month
# of each moving window.
# Trim down to a 10-year period so that the results are easier to see.

ggplot(clim1[clim1$year \%in\% 1991:2000,]) +
geom_path(aes(yr.mo, clim.var, color = win.len), na.rm = TRUE)
}
