% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_clim_gro_corr.R
\name{multi_clim_gro_corr}
\alias{multi_clim_gro_corr}
\title{Calculate multiple growth-climate relationships}
\usage{
multi_clim_gro_corr(
  rwl.group = NULL,
  clim.group = NULL,
  clim.var = NULL,
  group.IDs.df = NULL,
  group.var = NULL,
  gro.period.end = NULL,
  agg.fun = "mean",
  max.lag = 1,
  prewhiten = TRUE,
  corr.method = "spearman",
  hemisphere = NULL
)
}
\description{
This function is a workhorse function behind \code{\link{n_mon_corr}}. It splits and merges the
climate data and the rwl data, then calculates all the correlations between them.

Internal function for modendro.
}
