% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerkinsSwetnam96.R
\docType{data}
\name{PerkinsSwetnam96}
\alias{PerkinsSwetnam96}
\title{Perkins & Swetnam 1996 whitebark pine (Pinus albicaulis) data from central Idaho, USA}
\format{
\subsection{\code{PerkinsSwetnam96}}{

An rwl-data.frame with 1267 rows and 57 columns:
Like all rwl-format files, Years are contained as rownames and series IDs are column names
}
}
\source{
\url{https://www.ncei.noaa.gov/access/paleo-search/study/4110}

\url{https://www.ncei.noaa.gov/access/paleo-search/study/4109}

\url{https://www.ncei.noaa.gov/access/paleo-search/study/4111}

\url{https://www.ncei.noaa.gov/access/paleo-search/study/4108}

\url{https://doi.org/10.1139/x26-241}
}
\usage{
PerkinsSwetnam96
}
\description{
Tree-level mean ring widths (mm) calculated from Perkins & Swetnam's (1996) original raw data.
Downloaded from the International Tree-Ring Data Bank (ITRDB).
}
\details{
The tree series derive from 4 study sites in the mountains of central Idaho.
}
\keyword{datasets}
