% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_mon_corr.R
\name{plot_n_mon_corr}
\alias{plot_n_mon_corr}
\title{Summary plotting function for n_mon_corr output}
\usage{
plot_n_mon_corr(x = NULL)
}
\arguments{
\item{x}{the n_mon_cor output list object}
}
\value{
2 plots
}
\description{
Exploratory data analysis (EDA) function to compute correlations between tree ring data and a
monthly climate variable aggregated for every combination (lengths 1:12) of consecutive months
going back a specified number of years.
}
