% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ps.groupIDs}
\alias{ps.groupIDs}
\title{Site grouping data for Perkins & Swetnam whitebark pine collection, four Idaho sites, 1996
D. Perkins, T. Swetnam
Four sites (TWP, SDP, UPS, & RRR) from the original study are condensed into three
(TWP, SDP_UPS, & RRR). The SDP & UPS sites colocated - gridded climate estimates would not be
different.}
\format{
\subsection{\code{ps.groupIDs}}{

A data.frame with 68 rows and 2 columns
\describe{
\item{tree}{Tree identifier}
\item{site}{Site identifier}
...
}
}
}
\source{
Publication: \url{https://doi.org/10.1139/x26-241}
}
\usage{
ps.groupIDs
}
\description{
Site grouping data for Perkins & Swetnam whitebark pine collection, four Idaho sites, 1996
D. Perkins, T. Swetnam
Four sites (TWP, SDP, UPS, & RRR) from the original study are condensed into three
(TWP, SDP_UPS, & RRR). The SDP & UPS sites colocated - gridded climate estimates would not be
different.
}
\keyword{datasets}
