% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ps96}
\alias{ps96}
\title{Perkins & Swetnam whitebark pine ring width collection, four Idaho sites, 1996
D. Perkins, T. Swetnam
Four ITRDB collections are combined here and tree-level mean ring widths calculated.}
\format{
\subsection{\code{ps96}}{

An rwl data.frame with 1267 rows and 68 columns (68 tree-ring series spanning 1267 years)
}
}
\source{
Publication: \url{https://doi.org/10.1139/x26-241}
ITRDB collections:
\url{https://www.ncei.noaa.gov/access/paleo-search/study/4108}
\url{https://www.ncei.noaa.gov/access/paleo-search/study/4108}
\url{https://www.ncei.noaa.gov/access/paleo-search/study/4110}
\url{https://www.ncei.noaa.gov/access/paleo-search/study/4111}
}
\usage{
ps96
}
\description{
Perkins & Swetnam whitebark pine ring width collection, four Idaho sites, 1996
D. Perkins, T. Swetnam
Four ITRDB collections are combined here and tree-level mean ring widths calculated.
}
\keyword{datasets}
