% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSgroupIDs.R
\docType{data}
\name{PSgroupIDs}
\alias{PSgroupIDs}
\title{Data.frame containing grouping data for tree-ring series (tree IDs) from Perkins & Swetnam 1996
whitebark pine (Pinus albicaulis) data from central Idaho, USA. This is an example data.frame
for demonstrating the use of grouped data in \code{\link{n_mon_corr}}. PSgroupIDs is fed to the
\code{group.IDs.df} argument, and \code{"site"} is fed to the \code{group.var} argument. To be used when
\code{idPRISMgroup} is used as the climate data.}
\format{
\subsection{\code{PSgroupIDs}}{

A data.frame with 68 rows and 2 columns:
\describe{
\item{series}{The series ID code - in this case the tree ID code}
\item{site}{Site code corresponding to the Perkins & Swetnam study. SDP & UPS are combined as
coordinates were the same in the ITRDB}
...
}
}
}
\source{
\url{https://doi.org/10.1139/x26-241}
}
\usage{
PSgroupIDs
}
\description{
Data.frame containing grouping data for tree-ring series (tree IDs) from Perkins & Swetnam 1996
whitebark pine (Pinus albicaulis) data from central Idaho, USA. This is an example data.frame
for demonstrating the use of grouped data in \code{\link{n_mon_corr}}. PSgroupIDs is fed to the
\code{group.IDs.df} argument, and \code{"site"} is fed to the \code{group.var} argument. To be used when
\code{idPRISMgroup} is used as the climate data.
}
\keyword{datasets}
