% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwl_longer.R
\name{rwl_longer}
\alias{rwl_longer}
\title{Convert rwl-format dataframes/matrices to long format}
\usage{
rwl_longer(rwl = NULL, dat.name = "rw", omit.NAs = TRUE)
}
\arguments{
\item{rwl}{A rwl-type data.frame (e.g., read in by \code{\link[dplR]{read.rwl}}). Essentially a data.frame with columns names as series IDs and years as rownames.}

\item{dat.name}{Character vector of length 1 representing the column name you want for the the tree ring data. Default is "rw".}

\item{omit.NAs}{Logical vector indicating whether to keep NAs in the output or not. Default is TRUE. This can cause issues if missing rings are represented with NA.}
}
\value{
A data.frame with 3 columns
}
\description{
Simple function to convert rwl-format data (columns are series, rows are yeqrs) to long format
(3 columns: year, series, rw). This is useful for getting your data ready for plotting in ggplot or
for correlation analyses.
}
\examples{
library(dplR)
# Bristlecone pine tree ring collection from Campito mountain, White Mountains, CA
# Many 0 value rings in this collection
data("ca533")

ca533.long <- rwl_longer(ca533, dat.name = "rw.mm", omit.NAs = TRUE)
head(ca533.long)
}
