% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwl_replace_internal_NAs.R
\name{rwl_replace_internal_NAs}
\alias{rwl_replace_internal_NAs}
\title{Replace NA values when they occur within a rwl collection of tree-ring series}
\usage{
rwl_replace_internal_NAs(rwl = NULL, new.val = NULL)
}
\arguments{
\item{rwl}{A rwl-type data.frame (e.g., read in by \code{\link[dplR]{read.rwl}}). Essentially a
data.frame with columns names as series IDs and years as rownames.}

\item{new.val}{a numeric vector of length 1. The default is zero}
}
\description{
Simple utility function to replace NA values when they occur within a rwl. Chains of NA values
are common at the beginnings and ends of tree-ring series within rwl files, and by some
conventions missing or locally absent rings are marked with NA. This can cause problems for many
analyses.

This is a wrapper function for the simpler replace_internal_NAs that works on any single numeric
vector.

Warning: for the default of new.val = 0, make sure that the NAs do indeed represent missing or
locally absent rings!

Also, for some analyses, 0 values may cause issues. In these cases it might make sense to
provide some nominally small value, e.g., 0.001, 0.0001.
}
\examples{
data(PerkinsSwetnam96)
# replace some values with NAs
PS.NAs <- PerkinsSwetnam96
# Give at least one series an NA value
PS.NAs[sample(1:nrow(PS.NAs), 1),] <- NA

# Try to run a common dplR function that will throw an error if there are missing values in the
# middle of series
detrend(PS.NAs, method = "Spline")

PS.noNAs <- rwl_replace_internal_NAs(x = a.series, new.val = 0)

# No more error
detrend(PS.NAs, method = "Spline")
}
