% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xd_check.R
\name{xd_check}
\alias{xd_check}
\title{Statistical cross-dating checking using methods inspired by CooRecorder/CDendro and COFECHA.}
\usage{
xd_check(
  data = NULL,
  ref1 = NULL,
  ref2 = NULL,
  ref3 = NULL,
  std.method = "p2yrsL",
  max.offset = 10,
  p.thresh = 0.05,
  out.format = "rwl"
)
}
\arguments{
\item{data}{The collection of raw tree-ring series you wish to check. Can be class "rwl" or
"data.frame" (long format). This is required. If "data.frame" you need 3 columns: "series",
"year", & "rw".}

\item{ref1}{A reference collection of raw tree-ring series (NOT a chronology!). Can be class
"rwl" or "data.frame" (long format). Optional, but more is better here. If "data.frame" you need
3 columns: "series", "year", & "rw".}

\item{ref2}{A reference collection of raw tree-ring series (NOT a chronology!). Can be class
"rwl" or "data.frame" (long format). Optional, but more is better here. If "data.frame" you need
3 columns: "series", "year", & "rw".}

\item{ref3}{A reference collection of raw tree-ring series (NOT a chronology!). Can be class
"rwl" or "data.frame" (long format). Optional, but more is better here. If "data.frame" you need
3 columns: "series", "year", & "rw".}

\item{std.method}{The method of standardization/normalization to apply to the series and
references. Default (and currently only) option is "p2yrsL", same default as CooRecorder/CDendro.}

\item{max.offset}{The maximum offset (in years) you want to check in either direction from the
years as dated. Minimum is 5 years. No true maximum limit, but overlap between series and
references must be at least 20 years. Default is 10 years.}

\item{p.thresh}{The p-value threshold to determine if a correlation is significant. Default is
the common 0.05.}

\item{out.format}{The format of subsets of data based on the dating confidence ranking. Options
are "rwl" or "long".}
}
\value{
A list containing 1) the partitioned results based on relative confidence (itself a
4-element list), 2) the raw correlation results in a nested list for easy viewing in RStudio,
and 3) subsets of the data collection corresponding to the relative confidence (format specified
with \code{out.format}).
}
\description{
This function implements a custom statistical check on the dating of a collection of tree-ring
series. It is mostly inspired by operations in CDendro (Larsson & Larsson 2024) and COFECHA, but
does not do all of what either of those programs can do. \code{\link{xd_check}} does simple
whole-series cross-correlations between leave-one-out (LOO) chronologies and 1-3 reference
chronologies. You provide collections of raw tree-ring widths (not pre-made chronologies), and
xd_check will standardize/normalize all series and then generate the chronologies internally.
If you only provide the \code{data} collection, LOO comparisons are all that you'll get. Adding
external (and properly cross-dated!) reference collections adds a lot to the confidence of your
checks. The main output of the function is a summary based on how many references have
best and significant correlations with the series as dated and are ranked lowest, medium, and
highest confidence.

One potential use for this function is to generate "clean" references for subsequent runs of the
functions. You can do this by supplying the reference collection to the \code{data} argument and not
supplying any of the reference arguments (i.e., LOO only mode). Set the other arguments as you
please and then take the "Highest" subset as one of the \code{refn} arguments in the next run. Note
that this assumes that you have plenty of good series in your collection!
}
\details{
\code{\link{xd_check}} does a standardization/normalization on all individual series within the
\code{data} and \code{refn} collections, then it computes chronologies using Tukey's Biweight Robust mean
(via \code{\link[DescTools]{TukeyBiweight}}). The LOO method is common in dendrochronology as
the method to compute interseries correlations for a tree-ring collection. In LOO, each series
has its own chronology made up of the remaining series.
}
\examples{
# Examples to come
}
\references{
Larsson & Larsson (2024) \emph{CDendro and CooRecorder programs of the CDendro package},
Cybis Elektronik & Data AB. https://www.cybis.se/forfun/dendro/index.htm
}
