% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_detect.R
\name{ci_detect}
\alias{ci_detect}
\title{Curve intervention detection - statistical identification and removal of disturbances in tree ring series}
\usage{
ci_detect(
  rwl,
  detrend.method = "Mean",
  nyrs = NULL,
  min.win = 9,
  max.win = 30,
  var.type = "s_bi",
  thresh = 3.29,
  out.span = 1.25,
  max.iter = 10,
  add.recent.rwi = TRUE
)
}
\arguments{
\item{rwl}{A rwl object (read in by dplR's \code{read.rwl()}). Essentially a data.frame with columns names as series IDs and years as rownames.}

\item{detrend.method}{A character vector specifying the detrending method to use. Passes to \code{dplR::detrend()}. Default is "AgeDepSpline".}

\item{nyrs}{A numeric vector that determines the flexibility of the "AgeDepSpline" or the "Spline" detrending methods. Passes to \code{dplR::detrend()}. The default is 50 years.}

\item{min.win}{The minimum disturbance length in years (i.e., a moving window) to search for. The default is 9.}

\item{max.win}{The maximum disturbance length in years (i.e., a moving window) to search for. The default is the smallest of 30 years or 1/3 of the series length.}

\item{thresh}{The disturbance detection threshold, corresponding to the number of deviations from the robust mean. The default is 3.29, following Druckenbrod et al. 2013.}

\item{out.span}{Parameter to determine the wiggliness of the loess splines fit to disturbances periods. Higher numbers = less wiggles. Passes to \code{base::loess()}. The default is 1.25}

\item{max.iter}{The maximum number of iterations to run the disturbance detection and removal processes. The default is 10}
}
\value{
A 3-element list containing the "disturbance-free" series, the disturbance index, and a record of the outlier
detection and removal iterations.
}
\description{
This function implements an analysis based on the curve intervention detection techniques described by Druckenbrod et al. 2013,
Rydval et al. 2016, and Rydval et al. 2018. The modendro implementation differs from the version described by these authors. See Details for more information.
The basic motivations for performing this analysis are to identify, quantify, & remove abrupt changes in radial growth. You might use
this information to help make inferences about forest stand dynamics, insect outbreaks, or isolate a stronger climate signal, depending on your dendroecological or dendroclimatological tendencies.
}
\details{
Intervention detection is a statistical time series approach to identify and remove abrupt changes in radial growth. The "curve" part
of curve intervention detection describes the type of line fitted to each period that is identified as a disturbance.
}
\examples{
library(dplR)
data("co021")
# before
ci_detect(co021, max.iter = 5)
}
\references{
Druckenbrod, D. L., N. Pederson, J. Rentch, and E. R. Cook. 2013. A comparison of times series approaches for dendroecological reconstructions of past canopy disturbance events. Forest Ecology and Management 302:23–33.
Rydval, M., D. Druckenbrod, K. J. Anchukaitis, and R. Wilson. 2016. Detection and removal of disturbance trends in tree-ring series for dendroclimatology. Canadian Journal of Forest Research 401:387–401.
Rydval, M., D. L. Druckenbrod, M. Svoboda, V. Trotsiuk, P. Janda, M. Mikoláš, V. Čada, R. Bače, M. Teodosiu, and R. Wilson. 2018. Influence of sampling and disturbance history on climatic sensitivity of temperature limited conifers. The Holocene.
}
