% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_detrend.R
\name{d_detrend}
\alias{d_detrend}
\title{Disturbance detrending of tree-ring series based on Druckenbrod et al. 2024}
\usage{
d_detrend(
  data = NULL,
  win.len = 15,
  pgc.thresh = 50,
  d.detrend.method = "AgeDepSpline",
  detrend.method = "AgeDepSpline",
  nyrs = c(10, 25),
  event.type = "release",
  standardize = TRUE
)
}
\arguments{
\item{data}{The collection of raw tree-ring series you wish to disturbance-detrend. Can be class
\code{"rwl"} (e.g., read in by \code{\link[dplR]{read.rwl}}) or \code{"data.frame"} (long format). If
\code{"data.frame"} you need at least these 3 columns: \code{"series"}, \code{"year"}, & \code{"rw"}.}

\item{win.len}{The window length, in years, for calculating percent growth change from moving
averages of raw ring width. Default is 15 years.}

\item{pgc.thresh}{Percent growth change (pgc) threshold for detection of disturbance events.
Default is 50\%.}

\item{d.detrend.method}{Which method to use for detrending individual disturbance events. Options
are \code{"AgeDepSpline"}, \code{"Spline"}, \code{"Friedman"}, \code{"Hugershoff"}, \code{"Linear"}, and "\verb{Mean"}. Default
is \code{"AgeDepSpline"}. See details for more information.}

\item{detrend.method}{Which method to use for the final age detrending on each series. Options are
all the methods available in \code{\link[dplR]{detrend.series}}. Default is "AgeDepSpline".}

\item{nyrs}{Two numbers (in years) setting flexibility of splines for \code{d.detrend.method} and
\code{detrend.method}, respectively. For "AgeDepSpline", nyrs determines the initial flexibility. For
regular "Spline" the flexibility is constant throughout. Defaults are c(15, 25) years.}

\item{event.type}{The type of disturbance events to detrend. Options are "release", "suppression",
or "both".}

\item{standardize}{Standardize the final detrended series by dividing out the age trend curve?
Default is TRUE. If FALSE, the difference is taken instead, resulting in residual series.}
}
\value{
A list containing 3 data.frames 1) the disturbance detrended ring width results including
the percent growth change series and intermediate steps ("PGC"), 2) details about the specific
disturbance events detected, and 3) details on the disturbance event trend fitting and removal
process.

All data.frames are in "long-format" with series as a categorical variable.
}
\description{
\code{\link{d_detrend}} implements the disturbance detrending algorithm of Druckenbrod et al.
2024. This is an updated version of the same type of analysis as \code{\link{ci_detect}} - the
step-by-step identification and removal of disturbance events tree-ring series from closed-canopy
forests. The basic assumption is that rapid changes in growth are likely driven by stand
dynamics, not climate.

This newer method combines Nowacki & Abrams's (1997) radial growth averaging to detect
disturbances with the intervention detection techniques of Druckenbrod et al. 2013,
Rydval et al. 2016, and Rydval et al. 2018 to fit curves to the disturbance event periods and
remove them from tree-ring series. By default, events are fit with age-dependent splines
(Melvin et al. 2007).

The \code{modendro} implementation includes additional flexibility to tune parameters and try
different curve fitting methods, and also to allow the algorithm to detect and remove growth
suppression events in addition to growth releases.
}
\details{
Details to come
}
\examples{
Examples to come
}
\references{
Druckenbrod DL, Cook ER, Pederson N, Martin-Benito D. 2024. Detrending tree-ring widths in
closed-canopy forests for climate and disturbance history reconstructions.
\emph{Dendrochronologia}. \strong{85}:126195. https://doi.org/10.1016/j.dendro.2024.126195

Druckenbrod DL, Pederson N, Rentch J, Cook ER. 2013. A comparison of times series
approaches for dendroecological reconstructions of past canopy disturbance events.
\emph{Forest Ecology and Management}, \strong{302}:23–33.
https://doi.org/10.1016/j.foreco.2013.03.040

Rydval M, Druckenbrod D, Anchukaitis KJ, Wilson R. 2016. Detection and removal of
disturbance trends in tree-ring series for dendroclimatology.
\emph{Canadian Journal of Forest Research}, \strong{401}:387–401.
https://doi.org/10.1139/cjfr-2015-0366

Rydval M et al. 2018. Influence of sampling and disturbance history on climatic sensitivity of
temperature limited conifers. \emph{The Holocene}, \strong{28}(10):1574-1587.
https://doi.org/10.1177/0959683618782605

Melvin TM, Briffa KR, Nicolussi K, Grabner M. 2007. Time-varying-response smoothing.
\emph{Dendrochronologia}. \strong{25}(1):65–69. https://doi.org/10.1016/j.dendro.2007.01.004

Nowacki GJ, Abrams MD. 1997. Radial-Growth Averaging Criteria for Reconstruction Disturbance
Histories from Presettlement-Origin Oaks. \emph{Ecological Monographs}. \strong{67}(2):225.
https://doi.org/10.2307/2963514

Warren WG, MacWilliam SL. 1981. Test of a new method for removing the growth trend
from dendrochronological data.
\emph{Tree Ring Bulletin} \strong{41}:55–66.
}
\seealso{
\code{\link{plot_d_detrend}}, \code{\link{ci_detect}}
}
