% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_detrend.R
\name{d_detrend}
\alias{d_detrend}
\title{Disturbance detrending of tree-ring series based on Druckenbrod et al. 2024}
\usage{
d_detrend(
  data = NULL,
  win.len = 15,
  pgc.thresh = 50,
  d.detrend.method = "AgeDepSpline",
  detrend.method = "AgeDepSpline",
  nyrs = c(10, 30),
  event.type = "release"
)
}
\arguments{
\item{data}{The collection of raw tree-ring series you wish to disturbance-detrend. Can be class
\code{"rwl"} (e.g., read in by \code{\link[dplR]{read.rwl}}) or \code{"data.frame"} (long format). If
\code{"data.frame"} you need at least these 3 columns: \code{"series"}, \code{"year"}, & \code{"rw"}.}

\item{win.len}{The window length, in years, for calculating percent growth change from moving
averages of raw ring width. Default is 15 years.}

\item{pgc.thresh}{Percent growth change (pgc) threshold for detection of disturbance events.
Default is 50\%.}

\item{d.detrend.method}{Which method to use for detrending individual disturbance events. Options
are \code{"AgeDepSpline"}, \code{"Spline"}, \code{"Friedman"}, \code{"Hugershoff"}, \code{"Linear"}, and "\verb{Mean"}. Default
is \code{"AgeDepSpline"}. See details for more information.}

\item{detrend.method}{Which method to use for the final age detrending on each series. Options are
all the methods available in \code{\link[dplR]{detrend.series}} except for "Ar". Default is
"AgeDepSpline".}

\item{nyrs}{Two numbers (in years) setting flexibility of splines for \code{d.detrend.method} and
\code{detrend.method}, respectively. For "AgeDepSpline", nyrs determines the initial flexibility. For
regular "Spline" the flexibility is constant throughout. Defaults are c(10, 25) years.}

\item{event.type}{The type of disturbance events to detrend. Options are "release", "suppression",
or "both". Default is "release".}
}
\value{
A list containing 3 data.frames 1) the disturbance detrended ring width results including
the percent growth change series and intermediate steps ("PGC"), 2) details about the specific
disturbance events detected, and 3) details on the disturbance event trend fitting and removal
process.

All data.frames are in "long-format" with \code{series} as a categorical variable.
}
\description{
\code{\link{d_detrend}} implements the disturbance detrending algorithm of Druckenbrod et al.
(2024). This is an updated version of the same type of analysis as \code{\link{ci_detect}} - the
step-by-step identification and removal of disturbance events tree-ring series from closed-canopy
forests. The basic assumption is that rapid changes in growth are likely driven by stand
dynamics, not climate. \code{\link{d_detrend}} was adapted from MatLab code provided by
D. Druckenbrod.

This method combines Nowacki & Abrams's (1997) radial growth averaging to detect
disturbances with the intervention detection techniques of Druckenbrod et al. (2013),
Rydval et al. (2016), and Rydval et al. (2018) to fit curves to the disturbance event periods and
remove them from tree-ring series. By default, events are fit with age-dependent splines
(Melvin et al. 2007).

The \code{modendro} implementation includes additional flexibility to try different curve fitting
methods for disturbance events and final age detrending, and to allow the algorithm to detect and
remove growth suppression events in addition to growth releases. The default settings match the
parameters in Druckenbrod et al. (2024).
}
\details{
The concept of separately detrending disturbances derives from Cook's (1987) linear aggregate
model. Although Cook's model distinguishes between endogenous and exogenous disturbances, in
practice this is not possible without prior knowledge. Instead, disturbances are defined
quantitatively.

In \code{\link{d_detrend}} the percentage growth change (PGC) is calculated from
radial growth averaging (Nowacki & Abrams 1997) - i.e., the percentage change between adjacent (1
year of separation) \code{win.len} moving averages of ring width. Periods of PGC that are greater
than or equal to \code{abs(pgc.thresh)} are identified as disturbance events. Within the event
periods, the maximum PGC value is selected as the start year of the event. In this \code{modendro}
implementation, events that occur within the earliest 2*win.len years of the original tree-ring
series are ignored. This is to avoid treating juvenile growth trends as disturbance events. Also
unique the \code{modendro} version, you can control which type of event is detected & detrended with
the \code{event.type} argument.

Once events are identified, we power- or log-transform (see \code{\link{pwr_t_rwl}}) the whole
series based on the variance-mean (i.e., spread-level) relationship unique to each series, then
detrend them individually starting with the most recent event, moving backward in time until all
events are detrended. Event detrending involves extracting the subset of the transformed series
from the estimated start year (max PGC) to the end of the series. A line is then fit to the
subset, subtracted from the transformed values, and the 1st value of the original
power-transformed event subset is added back. Each event is then determined to be "sustained" or
"transient". If the detrended event values stay below/above the original values for the remainder
of the series, the event is sustained. If the detrended event values ever cross the original
values, the event is transient. Transient events are only detrended for the period between event
start and when they cross the original values.

You can choose the type of line used to detrend each event through \code{d.detrend.method}. The
default \code{"AgeDepSpline"} is the recommended method. You can adjust the initial flexibility of the
spline through the 1st element of \code{nyrs}. You may want to adjust depending on your goals/system.
Other methods are available as well to allow further exploration of techniques. Methods \code{Spline},
\code{Friedman}, \code{Linear}, and \code{Mean} are fairly straight-forward and follow the descriptions in
\code{\link[dplR]{detrend.series}}. Method \code{Hugershoff} is a custom implementation and follows
the "simple increment function" described by Warren & MacWilliam (1981), with the b parameter set
to 1. The curve is fit via \code{\link[stats]{nls}}. If the fit fails, the event is fit with
\code{"AgeDepSpline"}.

After all events are detrended, the long-term (whole series) age trend lines are fit to the
transformed series. Under the hood, this is just a call to \code{\link[dplR]{detrend.series}}.
Residual series are extracted here. You can elect to skip this age-detrending step by entering
\code{"none"} to the \code{detrend.method} argument, returning just the disturbance detrended series.

Finally, the power transformation is reversed for the disturbance detrended series and the fitted
age trend line, putting both in the original ring width units. Note that because the power
used was based on the empirical variance-mean relationship for the series, the variances of the
detrended events are also adjusted. In this way, the disturbance detrending is not a simple
subtraction. The last step is to compute the standardized ring width index by dividing the
reverse-transformed disturbance-detrended series by the reverse-transformed age trend.

For more details, see Druckenbrod et al. (2024).

I highly recommend you look at the results! Convenient visualizations of the
\code{\link{d_detrend}} process are available by using \code{\link{plot_d_detrend}}.

Details of output data.frames:
}
\section{"PGC"}{
\describe{
\item{series}{Series ID}
\item{year}{Year}
\item{rw}{Ring widths}
\item{pt.rw}{Transformed ring widths}
\item{optimal.pwr}{Estimated optimal power of transformation}
\item{action}{Type of transformation applied to series}
\item{win.len}{Moving average window length (constant for all series)}
\item{mav.Ra.M1}{Right-aligned moving average ring width}
\item{mav.La.M2}{Left-aligned moving average ring width}
\item{pgc}{Percent growth change}
\item{pt.rw.ddtrd}{Transformed dist.-detrended ring widths}
\item{message}{Information about ... Usually NA}
\item{pt.rw.ddtrd.resid}{pt.rw.ddtrd - pt.rw.ddtrd.At}
\item{pt.rw.ddtrd.At}{Fitted age trend (At) to transformed dist.-detrended ring widths}
\item{detrend.method}{Method used to fit At}
\item{rw.ddtrd}{Back-transformed dist.-detrended ring widths}
\item{rw.ddtrd.At}{Back-transformed fitted age trend (At)}
\item{Dt}{Back-transformed disturbacne signal}
\item{rw.ddtrd.index}{rw.ddtrd / rw.ddtrd.At}
...
}
}

\section{"Events"}{
\describe{
\item{series}{Series ID}
\item{max.ind}{Index position of max percentage growth change}
\item{max.val}{Absolute value of max percentage growth change}
\item{pgc.thresh}{User-specified percent growth change threshold}
\item{year}{Estimated first year of disturbance event}
\item{message}{Summary of distubance events detected for each series}
...
}
}

\section{"Dist. detrending"}{
\describe{
\item{series}{Series ID}
\item{year}{Year}
\item{pt.rw.i}{Transformed ring widths for this iteration}
\item{curve}{Disturbance curve fit in this iteration}
\item{pt.rw.ddtrd.i}{Transformed ring widths with disturbance curve removed in this iteration}
\item{iter}{Iteration ID}
\item{eventID}{Estimated first year of disturbance event}
\item{event.type}{release or suppression}
\item{event.dur}{sustained or transient}
\item{d.detrend.method}{Curve-fitting method for disturbance events}
...
}
}

\examples{

# Missouri post oak example from Druckenbrod et al. 2024

# Load Missouri post oak ring widths
data(mo024)

# Run d_detrend - use the same settings as in Druckenbrod et al. 2024
mo024.ddtrd <- d_detrend(data = mo024,
                         win.len = 15,
                         pgc.thresh = 50,
                         d.detrend.method = "AgeDepSpline",
                         detrend.method = "AgeDepSpline",
                         nyrs = c(10, 30),
                         event.type = "release")

# Generate plots using plot_d_detrend (output stored in a list)
mo024.ddtrd.plots <- plot_d_detrend(mo024.ddtrd)

# look at plots for one series - note that there are 3 plots for each series - in RStudio click
# the back arrow in the plot viewer to see all of them
mo024.ddtrd.plots$DEM01C

# Run again but with event.type = "both" to detrend both releases and suppressions
mo024.ddtrd.both <- d_detrend(data = mo024,
                              win.len = 15,
                              pgc.thresh = 50,
                              d.detrend.method = "AgeDepSpline",
                              detrend.method = "AgeDepSpline",
                              nyrs = c(10, 30),
                              event.type = "both")

# Summary look at the 3 output data.frames in the list
sapply(mo024.ddtrd, head)

# Generate plots using plot_d_detrend (output stored in a list)
mo024.ddtrd.both.plots <- plot_d_detrend(mo024.ddtrd.both)

# look at plots for one series
mo024.ddtrd.both.plots$DEM01C
}
\references{
Cook ER. 1987. The Decomposition of Tree-Ring Series for Environmental Studies.
\emph{Tree-Ring Bulletin}. \strong{47}:37–59

Druckenbrod DL, Cook ER, Pederson N, Martin-Benito D. 2024. Detrending tree-ring widths in
closed-canopy forests for climate and disturbance history reconstructions.
\emph{Dendrochronologia}. \strong{85}:126195. https://doi.org/10.1016/j.dendro.2024.126195

Druckenbrod DL, Pederson N, Rentch J, Cook ER. 2013. A comparison of times series
approaches for dendroecological reconstructions of past canopy disturbance events.
\emph{Forest Ecology and Management}, \strong{302}:23–33.
https://doi.org/10.1016/j.foreco.2013.03.040

Rydval M, Druckenbrod D, Anchukaitis KJ, Wilson R. 2016. Detection and removal of
disturbance trends in tree-ring series for dendroclimatology.
\emph{Canadian Journal of Forest Research}, \strong{401}:387–401.
https://doi.org/10.1139/cjfr-2015-0366

Rydval M et al. 2018. Influence of sampling and disturbance history on climatic sensitivity of
temperature limited conifers. \emph{The Holocene}, \strong{28}(10):1574-1587.
https://doi.org/10.1177/0959683618782605

Melvin TM, Briffa KR, Nicolussi K, Grabner M. 2007. Time-varying-response smoothing.
\emph{Dendrochronologia}. \strong{25}(1):65–69. https://doi.org/10.1016/j.dendro.2007.01.004

Nowacki GJ, Abrams MD. 1997. Radial-Growth Averaging Criteria for Reconstruction Disturbance
Histories from Presettlement-Origin Oaks. \emph{Ecological Monographs}. \strong{67}(2):225.
https://doi.org/10.2307/2963514

Warren WG, MacWilliam SL. 1981. Test of a new method for removing the growth trend
from dendrochronological data.
\emph{Tree Ring Bulletin} \strong{41}:55–66.
}
\seealso{
\code{\link{plot_d_detrend}}, \code{\link{ci_detect}}
}
