% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_opt_pwr.R
\name{find_opt_pwr}
\alias{find_opt_pwr}
\title{Find the optimal power for transforming a tree ring series to approximate homoscedasticity}
\usage{
find_opt_pwr(rwl)
}
\arguments{
\item{rwl}{A rwl object (read in by dplR's \code{read.rwl()}). Essentially a data.frame with columns names as series IDs and years as rownames.}
}
\value{
A named numeric vector with the series IDs (colnames) and the estimated optimal power of transformation
}
\description{
Use a spread versus level method to estimate the optimal power of transformation for a tree ring series
}
\details{
This function estimates the optimal power of transformation via the local spread versus level relationship of each series,
where local spread is defined as the absolute value of the first
differences, S, (|rwt - rwt-1|) and the local level is the arithmetic mean of each pair of adjacent values, M, (rwt + rwt-1)/2.
The spread versus level relationship is then modeled in a simple linear regression as log10(S) ~ log10(M). The optimal power of transformation
is then estimated as p = |1-slope| of this regression model.
See ?cp_detrend and Cook and Peters (1997) for more details.

Note that this is a very simple function that only implements the spread vs. level estimation
of an optimal power for making a tree ring series ± homoscedastic. This function does not select which of these estimated
powers to use, that is performed in the \code{pwr_t_rwl()} function. See ?pwr_t_rwl for more details.
}
\examples{
library(dplR)
data("ca533")
find_opt_pwr(rwl = ca533)
}
\references{
Cook, E. R., and Peters, K. (1997) Calculating unbiased tree-ring indices for the study of climatic and environmental change.
\emph{The Holocene}, \strong{7}(3), 361-370.
}
\seealso{
\code{\link[=pwr_t_rwl(), cp_detrend]{pwr_t_rwl(), cp_detrend()}}
}
