% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_mon_corr.R
\name{n_mon_corr}
\alias{n_mon_corr}
\title{Flexible monthly aggregate growth-climate cross correlations for exploratory data analysis}
\usage{
n_mon_corr(
  rwl = NULL,
  clim = NULL,
  clim.var = NULL,
  group.IDs.df = NULL,
  group.var = NULL,
  gro.period.end = NULL,
  agg.fun = "mean",
  max.lag = 1,
  prewhiten = TRUE,
  corr.method = "spearman",
  hemisphere = NULL,
  make.plot = TRUE
)
}
\arguments{
\item{rwl}{A rwl-type data.frame (e.g., read in by \code{\link[dplR]{read.rwl}}). Essentially a
data.frame with columns names as series IDs and years as rownames.}

\item{clim}{a \code{data.frame} with at least 3 columns: year, month (numeric), and a
climate variable.}

\item{clim.var}{character vector - the colname of the climate variable of interest in the \code{clim}
data.frame.}

\item{group.IDs.df}{an optional data.frame with 2 columns: "series", representing the names of
the tree-ring series (and matching the colnames of rwl) and a group.var (name is your
specification), representing a grouping (e.g., site, plot) variable.}

\item{group.var}{a character vector specifying a grouping variable (e.g., site, plot).
Must exist and be identical in clim and group.ID.df.}

\item{gro.period.end}{the last month in which you expect growth to occur for your study species
in your study region. Not crucial in this version - only draws a line in the output plot.}

\item{agg.fun}{character vector specifying the function to use for aggregating monthly
climate combinations. Options are "mean" or "sum", e.g., for temperature or precipitation data,
respectively. Default is "mean".}

\item{max.lag}{numeric vector specifying how many years of lag to calculate calculations for.
Default (and minimum) is 1 year.}

\item{prewhiten}{logical vector specifying whether or not to convert tree ring & climate time
series to ARIMA residuals (aka "prewhitening"). A "best fit" ARIMA model is automatically
selected using \code{\link[forecast]{auto.arima}}.
This removes most autocorrelation in a time series, leaving only the high-frequency variation.
This is common practice before using standard methods for cross-correlations. Default is TRUE.}

\item{corr.method}{character vector specifying which correlation method to use. Options are
\code{c("spearman", "kendall", "pearson")}. Default is \code{"spearman"} using the
\code{\link[corTESTsrd]{corTESTsrd}} function (also used for \code{corr.method = "kendall"}). This
method reduces the type I error rate associated with autocorrelated series. CAUTION: Currently
\code{corr.method = "pearson"} doesn't make any adjustments for autocorrelation. See Details below.}

\item{hemisphere}{a character vector specifying which hemisphere your tree ring data - &
climate data - comes from ("N" or "S"). Conventions for assigning growth years - and thus}

\item{make.plot}{logical vector indicating whether or not to produce a plot. Default is TRUE.
You will get a warning if you have < 10 tree-ring series. Ideally you have > 50.}
}
\value{
A 2-4 element list containing data.frames of the correlation results, the moving-window
prewhitened tree-ring data (if prewhiten = TRUE), and the default plot.
}
\description{
Exploratory data analysis (EDA) function to compute correlations between tree ring data and a
monthly climate variable aggregated for every combination (lengths 1:12) of consecutive months
going back a specified number of years.

Compared to methods that force rigidly-defined seasons of a fixed length, this approach should
facilitate discovery of potentially more meaningful growth-climate relationships.

Fair warning: this is a basic function that will accept any tree ring data and climate data in
the proper format. It is the user's responsibility to make sure that your data is appropriate
to the analyses.
}
\details{
Exploring a wide range of plausible grrowth-climate relationships can be a useful first step once
you have a collection of cross-dated tree ring series.

The default correlation test method is Spearman rank correlation. This will be ±equivalent
to Pearson for linear relationships, but will also capture any non-linear relationships. As an
additional precaution, \code{\link{n_mon_corr}} uses the \code{\link[corTESTsrd]{corTESTsrd}}
method from Lun et al. (2022) to reduce the type I error rate  (i.e., you think there is a
significant cross-correlation when there isn't one), which is inflated if there is
autocorrelation in your tree-ring or climate series. When autocorrelation is absent, the method
gives similar results as the classical significance test. For these reasons,
\code{\link{n_mon_corr}} always uses the Lun et al. method for Spearman and Kendall rank
correlations. This is done here as a precaution but also as an allowance so that you can examine
relationships between lower-frequency (which are autocorrelated) elements of tree-ring and
climate series as well as between the more common high-frequency signals (e.g., "prewhitened"
series). You may use Pearson correlations in \code{\link{n_mon_corr}}, but be forewarned that
there is no adjustment for autocorrelation in the Pearson correlation tests! Therefore, it
is general best to use \code{corr.method = "pearson"} for comparison only, rather than actual results.

A note on tree ring analyses based in the Southern hemisphere:
\code{\link{n_mon_corr}} is designed to work in both the Northern and Southern hemispheres.
Hemisphere matters for tree ring growth-climate relationships because tree ring formation in the
Southern hemisphere typically spans two calendar years (e.g., starting in Nov 2000 and ending
in Mar of 2001). It was Schulman's (1956) protocol to assign the earlier calendar year to the
tree rings in the Southern hemisphere, i.e., the calendar year in which growth began.
\code{\link{n_mon_corr}} assumes your data follows this standard as well. This has implications
for how the climate data is aligned with the treering data. In this version, the way this is
handled is that a "lag +1" year is added to suite of correlations so that the moving windows of
consecutive months can extend through the growing season (and into the next calendar year).
}
\examples{
# Make some synthetic stand-in tree ring data
rw.ex <- data.frame(year = 1:50, rw.mm = runif(50, 0.1, 2))


# Make some synthetic stand-in climate data
mo <- 1:12
x <- seq(-4, 4, length.out = 12)
gauss.curv <- \(x) {(10/sqrt(2*pi*1.6))*exp(-((x^2)/(2*1.6^2)))}
clim.mo <- data.frame(month = mo, clim.var = gauss.curv(x))
clim.list <- vector("list", length = 50)
for (y in seq_along(clim.list)) {
  clim.y <- clim.mo
  clim.y[,"clim.var"] <- clim.y[,"clim.var"] + runif(1, min = 0, max = 4)
  clim.y$year <- y
  clim.list[[y]] <- clim.y
}
clim <- do.call("rbind", clim.list)
n_mon_corr.out <- n_mon_corr(rw = rw.ex,
rw.col = "rw.mm",
clim = clim,
clim.var = "clim.var",
rel.per.begin = 3,
hemisphere = "S",
rw.name = "Synthetic RW")

# Take a look at the output data frames
head(n_mon_corr.out[["Correlation results"]])
head(n_mon_corr.out[["Correlation data"]])

## Method for applying n_mon_corr to each series (or tree) in a rwl file,
# consistent with tree-level analyses.

# Make some synthetic stand-in tree ring data (in rwl-type format)
rwl.ex <- matrix(nrow = 50, ncol = 10)
rownames(rwl.ex) <- 1:50
colnames(rwl.ex) <- 1:10
rwl.ex <- apply(rwl.ex, MARGIN = 2, FUN = \(x) runif(length(x), 0.1, 2)) |> as.data.frame()

# Convert rwl to long format using modendro's rwl_longer() function
rwl.ex.long <- modendro::rwl_longer(rwl.ex,
series.name = "tree",
dat.name = "rw.mm",
trim = TRUE)

# split the data.frame into a list based on ID
rwl.ex.long.list <- split(rwl.ex.long, f = rwl.ex.long$tree)

# Use mapply to run n_mon_corr for each "tree"
# Note that in our example, the clim data is the same for all series.
# You will need a different processes if you have climate and tree ring series grouped by site
# or plot.
# Also note that plots = FALSE and silent = TRUE so that these don't clog the plotting window
# and console.
n_mon_corr.out.list <- lapply(rwl.ex.long.list, FUN = \(x) {
n_mon_corr(rw = x,
rw.col = "rw.mm",
clim = clim,
clim.var = "clim.var",
rel.per.begin = 3,
hemisphere = "S",
plots = FALSE,
silent = TRUE)
})

# Outputs are the same as above, but nested in one more list dimension
head(n_mon_corr.out.list[[1]][["Correlation results"]])

# It might be helpful to rbind the individual tree outputs into data.frames

data.df <- lapply(n_mon_corr.out.list, FUN = \(x, y) {
x[["Correlation data"]]
}) |>
 do.call(what = "rbind")

head(data.df)

results.df <- mapply(FUN = \(x, y) {
 x[["Correlation results"]]$tree <- y
 x[["Correlation results"]]
}, x = n_mon_corr.out.list, y = names(n_mon_corr.out.list), SIMPLIFY = FALSE) |>
 do.call(what = "rbind") |> as.data.frame()

head(results.df)

# Now it is possible to do things like find out how many trees have significant correlations
# for each month combination (these are random series so results are not meaningful):
results.df$sig <- ifelse(results.df$p < 0.05, "Sig.","Not sig.")
results.agg <- aggregate(tree ~ months + sig, data = results.df, length)
results.agg[order(results.agg$tree, decreasing = TRUE),] |> head()
}
\references{
Schulman, E. (1956) \emph{Dendroclimatic changes in semiarid America},
University of Arizona Press.

Lun, D., S. Fischer, A. Viglione, and G. Blöschl. (2022). Significance testing of rank
cross-correlations between autocorrelated time series with short-range dependence,
\emph{Journal of Applied Statistics}:1–17.
}
