% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_mon_corr.R
\name{n_mon_corr}
\alias{n_mon_corr}
\title{Growth-climate correlations via moving windows}
\usage{
n_mon_corr(
  chrono = NULL,
  clim = NULL,
  var = NULL,
  site = NULL,
  chrono.col = "xxxstd",
  agg.fun = "mean",
  max.lag = 3,
  corr.method = "pearson",
  plots = TRUE
)
}
\arguments{
\item{chrono}{A chron object (such as that produced by dplR's \code{chron()}).}

\item{clim}{a \code{data.frame} with at least 3 columns: year, month (numeric), and a climate variable.}

\item{var}{character vector - the name of the climate variable of interest in the \code{clim} data.frame (i.e., the colname).}

\item{site}{character vector - the name of your site.}

\item{chrono.col}{A chron object (such as that produced by dplR's \code{chron()}).}

\item{agg.fun}{character vector specifying the function to use for aggregating monthly climate combinations.
Options are "mean" or "sum", e.g., for temperature or precipitation data, respectively. Default is "mean".}

\item{max.lag}{numeric vector specifying how many years of lag to calculate calculations for. Default is 3 years.}

\item{corr.method}{character vector specifying which correlation method to use. Passes to \code{cor.test()}.}

\item{plots}{logical vector indicating whether or not to produce plots.}
}
\value{
A named numeric vector with the series IDs (colnames) and the estimated optimal power of transformation
}
\description{
Function to take a tree ring chronology and a monthly climate variable to compute
monthly aggregate correlations for every combination of consecutive months for calendar years and water years (Oct-Sept)
The function computes the correlations for a specified number lagged years (ie, years before ring formation).
3 years is the default and is a sensible maximum for most analyses.
}
\details{
This needs work
}
\examples{
Will add some later

}
