% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_mon_corr.R
\name{n_mon_corr}
\alias{n_mon_corr}
\title{Flexible monthly aggregate growth-climate cross correlations for exploratory data analysis}
\usage{
n_mon_corr(
  rwl = NULL,
  clim = NULL,
  clim.var = NULL,
  common.years = NULL,
  agg.fun = "mean",
  max.win = 6,
  win.align = "left",
  max.lag = 1,
  hemisphere = NULL,
  prewhiten = TRUE,
  corr.method = "spearman",
  gro.period.end = NULL,
  group.IDs.df = NULL,
  group.var = NULL
)
}
\arguments{
\item{rwl}{A rwl-type data.frame (e.g., read in by \code{\link[dplR]{read.rwl}}). Essentially a
data.frame with columns names as series IDs and years as rownames.}

\item{clim}{a \code{data.frame} with at least 3 columns: year, month (numeric), and a
climate variable.}

\item{clim.var}{character vector - the colname of the climate variable of interest in the \code{clim}
data.frame.}

\item{common.years}{numeric vector - a sequence of years to subset the clim and rwl data before
running correlation analyses. Must be at least 25 years long and years must exist in clim and
rwl. This has advantages over subsetting the climate or tree-ring data a priori. See Details.}

\item{agg.fun}{character vector specifying the function to use for aggregating monthly
climate combinations. Options are "mean" or "sum", e.g., for temperature or precipitation data,
respectively. Default is "mean".}

\item{max.win}{integer vector specifying how long, in months, the longest (or widest) moving
window is. Values limited to between 2 and 12. Default is 6 months.}

\item{win.align}{the alignment of the moving windows. Options are "left" or "right". If "left",
month will indicate the starting month of each moving window and NAs will appear at the end of
the series. If "right", month will indicate the ending month of each moving window and NAs appear
at the beginning. The default is "left".}

\item{max.lag}{integer vector specifying how many years of lag to calculate calculations for.
Default (and minimum) is 1 year.}

\item{hemisphere}{a character vector specifying which hemisphere your tree ring data - &
climate data - comes from ("N" or "S"). Conventions for assigning growth years - and thus}

\item{prewhiten}{logical vector specifying whether or not to convert tree ring & climate time
series to ARIMA residuals (aka "prewhitening"). A "best fit" ARIMA model is automatically
selected using \code{\link[forecast]{auto.arima}}.
This removes most autocorrelation in a time series, leaving only the high-frequency variation.
This is common practice before using standard methods for cross-correlations. Default is TRUE.}

\item{corr.method}{character vector specifying which correlation method to use. Options are
\code{c("spearman", "kendall", "pearson")}. Default is \code{"spearman"} using the
\code{\link[corTESTsrd]{corTESTsrd}} function (also used for \code{corr.method = "kendall"}). This
method reduces the type I error rate associated with autocorrelated series. CAUTION: Currently
\code{corr.method = "pearson"} doesn't make any adjustments for autocorrelation. See Details below.}

\item{gro.period.end}{the last month in which you expect growth to occur for your study species
in your study region. Not crucial in this version - only draws a line in the output plot.}

\item{group.IDs.df}{an optional data.frame with 2 columns: "series", representing the names of
the tree-ring series (and matching the colnames of rwl) and a group.var (name is your
specification), representing a grouping (e.g., site, plot) variable.}

\item{group.var}{a character vector specifying a grouping variable (e.g., site, plot).
Must exist and be identical in clim and group.ID.df.}
}
\value{
A 2-4 element list containing data.frames of the correlation results, the moving-window
prewhitened tree-ring data (if prewhiten = TRUE), and the default plot.
}
\description{
Exploratory data analysis (EDA) function to compute correlations between tree ring data and a
monthly climate variable aggregated for every combination (lengths 1:12) of consecutive months
going back a specified number of years.

Compared to methods that force rigidly-defined seasons of a fixed length, this approach should
facilitate discovery of potentially more meaningful growth-climate relationships.

Fair warning: this is a basic function that will accept any tree ring data and climate data in
the proper format. It is the user's responsibility to make sure that your data is appropriate
to the analyses.
}
\details{
Exploring a wide range of plausible growth-climate relationships can be a useful first step once
you have a collection of cross-dated tree ring series.

The default correlation test method is Spearman rank correlation. This will be ±equivalent
to Pearson for linear relationships, but will also capture any non-linear relationships. As an
additional precaution, \code{\link{n_mon_corr}} uses the \code{\link[corTESTsrd]{corTESTsrd}}
method from Lun et al. (2022) to reduce the type I error rate  (i.e., you think there is a
significant cross-correlation when there isn't one), which is inflated if there is
autocorrelation in your tree-ring or climate series. When autocorrelation is absent, the method
gives similar results as the classical significance test. For these reasons,
\code{\link{n_mon_corr}} always uses the Lun et al. method for Spearman and Kendall rank
correlations. This is done here as a precaution but also as an allowance so that you can examine
relationships between lower-frequency (which are autocorrelated) elements of tree-ring and
climate series as well as between the more common high-frequency signals (e.g., "prewhitened"
series). You may use Pearson correlations in \code{\link{n_mon_corr}}, but be forewarned that
there is no adjustment for autocorrelation in the Pearson correlation tests! Therefore, it
is general best to use \code{corr.method = "pearson"} for comparison only, rather than actual results.

The \code{common.years} argument is a way to subset the climate and tree-ring data after the moving
windows and lags have been calculated. This results in lagged correlations that cover the same
number of years as the current year correlations. This will not be the case if you subset the
climate data before feeding it to \code{n_mon_corr}. E.g., if there is a 60 year overlap between a
climate dataset and a tree-ring series, the lags will erode years from the total n of the
correlations - a 1 year lag will be n-1, a 2 year lag will be n-2, etc. This is simply because of
the offsets. Specifying common.years solves this problem by setting up the lags before the data
is subset. The major caveat is that the climate data have to extend before \code{min(common.years)} in
order for this to work. Otherwise you get the eroding years issue.

A note on tree ring analyses based in the Southern hemisphere:
\code{\link{n_mon_corr}} is designed to work in both the Northern and Southern hemispheres.
Hemisphere matters for tree ring growth-climate relationships because tree ring formation in the
Southern hemisphere typically spans two calendar years (e.g., starting in Nov 2000 and ending
in Mar of 2001). It was Schulman's (1956) protocol to assign the earlier calendar year to the
tree rings in the Southern hemisphere, i.e., the calendar year in which growth began.
\code{\link{n_mon_corr}} assumes your data follows this standard as well. This has implications
for how the climate data is aligned with the treering data. In this version, the way this is
handled is that a "lag +1" year is added to suite of correlations so that the moving windows of
consecutive months can extend through the growing season (and into the next calendar year).
}
\examples{
## Bring in some real tree-ring and climate data
# Tree-ring data from Perkins & Swetnam 1996 (https://doi.org/10.1139/x26-241)
data(PerkinsSwetnam96)
# PRISM (https://www.prism.oregonstate.edu/) time series extracted for each site then averaged
# over all sites.
data(idPRISM)

# Monthly average temperature
PS_gro_Tavg <- n_mon_corr(rwl = PerkinsSwetnam96,
                         clim = idPRISM,
                         clim.var = "Tavg.C",
                         common.years = 1901:1990,
                         agg.fun = "mean",
                         max.win = 6,
                         win.align = "left",
                         max.lag = 2,
                         hemisphere = "N",
                         prewhiten = TRUE,
                         corr.method = "spearman",
                         gro.period.end = 9,
                         group.IDs.df = NULL,
                         group.var = NULL)
names(PS_gro_Tavg)
PS_gro_Tavg$`Results plots`
# April has the strongest signal (largest \% of significant correlations), and it is a negative
# relationship.


# Monthly total precipitation
# Note that agg.fun = "sum" for precipitation data.
PS_gro_PPT <- n_mon_corr(rwl = PerkinsSwetnam96,
                        clim = idPRISM,
                        clim.var = "PPT.mm",
                        common.years = 1901:1990,
                        max.win = 6,
                        win.align = "left",
                        agg.fun = "sum",
                        max.lag = 2,
                        hemisphere = "N",
                        prewhiten = TRUE,
                        corr.method = "spearman",
                        gro.period.end = 9,
                        group.IDs.df = NULL,
                        group.var = NULL)
names(PS_gro_PPT)
PS_gro_PPT$`Results plots`
# Strongest signal is Oct-Jan total precip.


## Demonstrate grouped data
data(idPRISMgroup)
data(PSgroupIDs)

PS_gro_Tavg_grouped <- n_mon_corr(rwl = PerkinsSwetnam96,
                                 clim = idPRISMgroup,
                                 clim.var = "Tavg.C",
                                 common.years = 1901:1990,
                                 max.win = 6,
                                 win.align = "left",
                                 agg.fun = "mean",
                                 max.lag = 2,
                                 hemisphere = "N",
                                 prewhiten = TRUE,
                                 corr.method = "spearman",
                                 gro.period.end = 9,
                                 group.IDs.df = PSgroupIDs,
                                 group.var = "site")
names(PS_gro_Tavg_grouped)
PS_gro_Tavg_grouped$`Results plots`
# Similar result, but not identical - now month = Apr with a 2-month moving window
# The climate data is slightly different for each site, so some differences in results are not
# surprising.
}
\references{
Schulman, E. (1956) \emph{Dendroclimatic changes in semiarid America},
University of Arizona Press.

Lun, D., S. Fischer, A. Viglione, and G. Blöschl. (2022). Significance testing of rank
cross-correlations between autocorrelated time series with short-range dependence,
\emph{Journal of Applied Statistics}:1–17.
}
