% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_det_rem.R
\name{out_det_rem}
\alias{out_det_rem}
\title{Supporting function for \code{ci_detect()} - statistical identification and removal of outliers in tree ring series}
\usage{
out_det_rem(
  rwi,
  min.win = 9,
  max.win = 30,
  var.type = "s_bi",
  thresh = 3.29,
  out.span = 1.25,
  add.recent.rwi = TRUE
)
}
\arguments{
\item{rwi}{A rwl object of detrended standardized series produced by \code{cp_detrend()}.}

\item{min.win}{The minimum outlier length in years (i.e., a moving window) to search for. The default is 9.}

\item{max.win}{The maximum outlier length in years (i.e., a moving window) to search for. The default is 30.}

\item{out.span}{Parameter to determine the wiggliness of the loess splines fit to outlier periods. Higher numbers = more wiggles. Passes to \code{stats::loess()}. The default is 1.}
}
\description{
This function implements the outlier detection and premoval steps for the curve intervention detection techniques described by Druckenbrod et al,
Rydval et al, and Rydval et al.
}
