% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ci_detect.R
\name{plot_ci_detect}
\alias{plot_ci_detect}
\title{Plot the \code{ci_detect()} processes}
\usage{
plot_ci_detect(ci_output)
}
\arguments{
\item{ci_output}{A list produced by the \code{ci_detect()} function}
}
\value{
A nested list of output plots illustrating the disturbance detection & removal iterations (1st list element)
for each tree ring series & the final result of the disturbance-free series after all iterations (2nd list element).
}
\description{
Takes the nested output list from \code{ci_detect()} and makes a list of plots that show the process for each series.
}
\details{
These plots are designed to give the user a window into how the \code{ci_detect()} process works.
}
