% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cp_detrend.R
\name{plot_cp_detrend}
\alias{plot_cp_detrend}
\title{Plot the cp_detrend process}
\usage{
plot_cp_detrend(cp_out)
}
\arguments{
\item{cp_out}{A list produced by the \code{\link{cp_detrend}} function}
}
\value{
A list of output plots showing the power transformation and detrending processes for each series.
}
\description{
Takes the output from \code{\link{cp_detrend}} and makes a list of plots that show the Cook & Peters (1997) process for each series.
}
\examples{

library(dplR)
data("ca533")
ca533_cp <- cp_detrend(ca533_cp, detrend.method = "AgeDepSpline")
ca533_cp_plots <- plot_cp_detrend(ca533_cp)
# View the first plot
ca533_cp_plots[[1]]

}
\references{
Cook, E. R., and Peters, K. (1997) Calculating unbiased tree-ring indices for the study of climatic and environmental change.
\emph{The Holocene}, \strong{7}(3), 361-370.
}
\seealso{
\code{\link{cp_detrend}}, \code{\link{pwr_t_rwl}}, \code{\link{find_opt_pwr}}
}
