% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cp_detrend.R
\name{plot_cp_detrend}
\alias{plot_cp_detrend}
\title{Plot the \code{cp_detrend()} process}
\usage{
plot_cp_detrend(cp_out)
}
\arguments{
\item{cp_out}{A list produced by the \code{cp_detrend()} function}
}
\value{
A list of output plots showing the power transformation and detrending processes for each series.
}
\description{
Takes the output from \code{cp_detrend()} and makes a list of plots that show the Cook & Peters (1997) process for each series.
}
\examples{

library(dplR)
data("ca533")
ca533_cp <- cp_detrend(ca533_cp, detrend.method = "AgeDepSpline")
ca533_cp_plots <- plot_cp_detrend(ca533_cp)
ca533_cp_plots[[1]]

}
\seealso{
\code{\link[=cp_detrend(), pwr_t_rwl(), find_opt_pwr]{cp_detrend(), pwr_t_rwl(), find_opt_pwr()}}
}
