% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_d_detrend.R
\name{plot_d_detrend}
\alias{plot_d_detrend}
\title{Plot the d_detrend processes & final results}
\usage{
plot_d_detrend(x = NULL)
}
\arguments{
\item{x}{A list produced by the \code{\link{d_detrend}} function}
}
\value{
A nested list of 3 output plots per series illustrating the disturbance detection process
(\code{"Detection plots"}), the curve fitting and removal steps for each event (\code{"Dist. detrending"})
& the final result of the disturbance-detrended series befre and after final age detrending
(\code{"Result plots"}).
}
\description{
Takes the nested output list from \code{\link{d_detrend}} and makes a list of plots that show
the disturbance detrending process for each series.
}
\details{
These plots are designed to illustrate how the \code{\link{d_detrend}} process works & visualize
the final results for each tree ring series. They are based on the plots in Druckenbrod et al.
(2024).
}
\examples{
examples to come
}
\seealso{
\code{\link{d_detrend}}
}
