% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_mon_corr.R
\name{plot_n_mon_corr}
\alias{plot_n_mon_corr}
\title{Summary plotting function for n_mon_corr output}
\usage{
plot_n_mon_corr(x = NULL)
}
\arguments{
\item{x}{the n_mon_corr output list object}
}
\value{
2 plots
}
\description{
Exploratory data analysis (EDA) function to compute correlations between tree ring data and a
monthly climate variable aggregated for every combination (lengths 1:12) of consecutive months
going back a specified number of years.

The results plots show 1) the percentage of tree-ring series that had statistically significant
correlations and 2) the mean correlation coefficient (ALL correlations, not just significant
ones) with all possible moving window combinations. Moving windows are represented by their start
month (the x-axis) and window length (represented by the lines of different colors). Each plot
is split by the direction of the relationships: positive (coef > 0) and negative (coef < 0).
The x-axis is extended to the left according to the number of lag years the user specifies (with
\code{max.lag}). Lag years are indicated by labeled rectangles just above the x-axis ticks.
"0" represents the current year.
}
