% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pos.R
\name{read_pos}
\alias{read_pos}
\title{Read in CooRecorder .pos files}
\usage{
read_pos(path = NULL)
}
\arguments{
\item{path}{file path to a single .pos file or a directory that may contain several .pos files.}
}
\value{
A list containing 2-3 data.frames for ring widths, attributes, and the original
coordinates (for single .pos files only).
}
\description{
This function reads in .pos files from CooRecorder (Cybis Elektronik & Data AB, Larsson &
Larsson). \code{read_pos} can handle a single file or a directory containing multiple .pos files.
}
\details{
If path is a directory the function will search all sub-directories for .pos files, thus
accomodating a range of directory structures. The output is slightly different if \code{path} leads to
a single file - the "raw" coordinates are also returned in this case.
The main outputs are a "Ring widths" data.frame containing the ring widths (whole ring, late wood
, and early wood, as applicable) and any year-specific point labels and an "Attributes"
data.frame containing the distance to pith, the outer data, the inner-most date, the radius (sum
of all ring widths plus the distance to pith), and the comment for whole series.
}
