% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pos.R
\name{read_pos}
\alias{read_pos}
\title{Read in CooRecorder .pos files}
\usage{
read_pos(path = NULL)
}
\arguments{
\item{path}{file path to a single .pos file or a directory that may contain several .pos files.}
}
\value{
A list containing 2-3 data.frames for ring widths, attributes, and the original
coordinates (for single .pos files only). The ring widths are in "long format", which is the more
flexible format for most analyses. If you need to export this data as a .rwl, for example, you
can use \code{\link{longer_rwl}} and dplR's \code{\link[dplR]{write.rwl}}.
}
\description{
This function reads in .pos files from CooRecorder (Cybis Elektronik & Data AB, Larsson &
Larsson). \code{read_pos} can handle a single file or a directory containing multiple .pos files.

Note that in the current implementation, the multiple .pos files mode can be pretty slow.

The motivation of this function is to replicate some of the basic operations of CDendro in R.
This allows the user to avoid saving tree-ring collections using some of the arcane tree-ring
file formats which don't allow users to take full advantage of the great features of CooRecorder,
such as point labels, comments, etc. It also calculates the ring widths with greater precision
(they are not rounded by default) than what CDendro exports.

A possible error in CooRecorder is that points can be saved out of order. Usually CooRecorder
will give you an "Erroneous point order" message, but it will save the file anyhow. This can
wreak havoc on determining ring widths. read_pos has a simple way of determining if there are
out of order points and it will warn you if it finds something but will read in the file anyway.
Go check those files in CooRecorder - some will be false positives, others will truly be messed
up.

For the forseeable future, \code{read_pos} will only work with .pos files from CooRecorder 7.8 or
greater - which is the earliest version (I think!) to include the actual pith coordinates in the
.pos files. There is also the stipulation that the file \emph{must} contain the outer year of the ring
width measurements. I'm not sure which version of CooRecorder started including this.
}
\details{
If path is a directory the function will search all sub-directories for .pos files, thus
accomodating a range of directory structures. The output is slightly different if \code{path} leads to
a single file - the "raw" coordinates are also returned in this case.
The main outputs are a "Ring widths" data.frame containing the ring widths (whole ring, late wood
, and early wood, as applicable) and any year-specific point labels and an "Attributes"
data.frame containing the distance to pith, the outer data, the inner-most date, the radius (sum
of all ring widths plus the distance to pith), and the comment for whole series.
}
\examples{
# example code

}
\references{
Larsson & Larsson (2023) \emph{CDendro and CooRecorder programs of the CDendro package},
Cybis Elektronik & Data AB. https://www.cybis.se/forfun/dendro/index.htm
}
\seealso{
\code{\link{longer_rwl}}, \code{\link{rwl_longer}}, \code{\link{p2yrsL}}
}
