% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwl_longer.R
\name{rwl_longer}
\alias{rwl_longer}
\title{Convert rwl-format dataframes/matrices to long format}
\usage{
rwl_longer(
  rwl = NULL,
  series.name = "series",
  dat.name = "rw",
  trim = TRUE,
  new.val.internal.na = NULL
)
}
\arguments{
\item{rwl}{A rwl-type data.frame (e.g., read in by \code{\link[dplR]{read.rwl}}). Essentially a
data.frame with columns names as series IDs and years as rownames.}

\item{series.name}{Character vector of length 1 for the column name you want for the series IDs.
Default is "series", but perhaps "tree" or "sample" is more appropriate for your work.}

\item{dat.name}{Character vector of length 1 for the column name you want for the the tree ring
data. Default is "rw".}

\item{trim}{Logical vector indicating whether to trim off NA sequences at the beginning or end
of individual series. Default is TRUE. Will not remove missing rings that are represented with
NA (i.e., NA values within the series).}

\item{new.val.internal.na}{Numeric vector giving a new value to replace NAs in the
middle of measurement series (i.e., some measurement protocols leave NAs for missing/locally
absent rings). E.g., \code{0} makes sense. If \code{NULL}, these values are left as NAs.}
}
\value{
A data.frame with 3 columns: 1) "year", 2) "series", 3) dat.name ("rw", "rw.mm",
"bai.mm" or whatever you name this)
}
\description{
Simple function to convert rwl-format data (columns are series, rows are years) to long format
(3 columns: year, series, rw). This is useful for getting your data ready for plotting in ggplot
or for correlation analyses.

\code{\link{rwl_longer}} doesn't tolerate replicate column names. Column names are series names,
after all. If you have replicate column names in your rwl, you will get an error message
that lists the replicated columns. Fix these and run \code{\link{rwl_longer}} again.
}
\examples{
library(dplR)
# Bristlecone pine tree ring collection from Campito mountain, White Mountains, CA
# Many 0 value rings in this collection
data("ca533")

ca533.long <- rwl_longer(rwl = ca533, dat.name = "rw.mm", trim = TRUE)
head(ca533.long)
}
